pyproject.toml
[tool.coverage.run]
omit = ["mcp_server/tests/*"]

[tool.coverage.report]
fail_under = 66
skip_empty = true

[tool.coverage.html]
directory = ".tests_coverage"

[tool.ruff]
preview = true
line-length = 120
fix = true
output-format = "grouped"

[tool.ruff.format]
exclude = [".git", ".venv", ".mypy_cache", ".tox", "__pycache__"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_unused_ignores = true
strict_equality = true
extra_checks = true
warn_unused_configs = true
warn_redundant_casts = true

[project]
name = "jenkins-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastmcp>=2.11.1",
  "python-jenkins>=1.8.3",
  "python-simple-logger>=2.0.16",
]


[project.scripts]
jenkins-mcp = "mcp_server.main:mcp.run"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=9.4.0",
]
tests = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.11.0",
    "pytest-asyncio>=0.21.0",
]
